+++ pkg/sql/colencoding/value_encoding.go
@@ -26,7 +26,7 @@ import (
 // the result to the idx'th position of the input exec.Vec.
 // See the analog in sqlbase/column_type_encoding.go.
 func DecodeTableValueToCol(
-	vec coldata.Vec, idx uint16, typ encoding.Type, dataOffset int, valTyp *types.T, b []byte,
+	vec coldata.Vec, idx uint64, typ encoding.Type, dataOffset int, valTyp *types.T, b []byte,
 ) ([]byte, error) {
 	// NULL is special because it is a valid value for any type.
 	if typ == encoding.Null {
@@ -50,7 +50,7 @@ func DecodeTableValueToCol(
 // If t is types.Bool, the value tag must be present, as its value is encoded in
 // the tag directly.
 // See the analog in sqlbase/column_type_encoding.go.
-func decodeUntaggedDatumToCol(vec coldata.Vec, idx uint16, t *types.T, buf []byte) ([]byte, error) {
+func decodeUntaggedDatumToCol(vec coldata.Vec, idx uint64, t *types.T, buf []byte) ([]byte, error) {
 	var err error
 	switch t.Family() {
 	case types.BoolFamily:
